+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes
 
-## 0.2.0 [2025/05/22] - AUSCERT Release
+## 0.2.0 [2025/05/22] - AUSCERT/SINCON Release
 
 **New Features:**